This commit includes these fixes
This commit fixes problems with Flylint, C/C++ Extension for VSCode, and
also adds settings.json and extensions.json so that they can be used in the future for configuration of this repository.

After I added some stuff in settings.json for the extension and set everything to use only Clang it went away.